pyproject.toml
[tool.poetry]
name = "tgbot-member-presence"
version = "0.1.0"
description = "Chat Member Presence Telegram Bot"
authors = ["Illia Ananich <illia.ananich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.3"
django-environ = "^0.4.5"
SQLAlchemy = "^1.3.23"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
